python/requirements.txt
opencv-python
mediapipe
pytesseract
numpy
pandas
scikit-learn
flask
flask
flask-cors
opencv-python
mediapipe
pytesseract
numpy
pandas
scikit-learn
PyPDF2
pdfplumber
python-docx
pdf2image
Pillow
pymupdf==1.23.8
scikit-learn==1.3.2
python-docx==1.1.0
langdetect==1.0.9
beautifulsoup4==4.12.2